Added missing RangedHeal to CreepOrderType. Changed IMarket.Credits from int to double. Added Name value variants to IMemoryObject.

+++ ScreepsDotNet.API/API/World/IMarket.cs
@@ -247,7 +247,7 @@ namespace ScreepsDotNet.API.World
         /// <summary>
         /// Your current credits balance.
         /// </summary>
-        int Credits { get; }
+        double Credits { get; }
 
         /// <summary>
         /// An array of the last 100 incoming transactions to your terminals.
